refactor: dto 패키지 구조 변경 및 setter 제거 (#168)
