+++ src/main/java/chungbazi/chungbazi_be/domain/user/converter/UserConverter.java
@@ -1,9 +1,7 @@
 package chungbazi.chungbazi_be.domain.user.converter;
 
-import chungbazi.chungbazi_be.domain.user.dto.UserResponseDTO;
+import chungbazi.chungbazi_be.domain.user.dto.response.UserResponseDTO;
 import chungbazi.chungbazi_be.domain.user.entity.User;
-import chungbazi.chungbazi_be.domain.user.entity.enums.RewardLevel;
-import java.util.List;
 
 public class UserConverter {
     public static UserResponseDTO.ProfileDto toProfileDto(User user) {
